requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.1.1
Pillow==6.2.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1